Several fixes
git-svn-id: svn://svn.dolezel.info/fatrat/trunk/fatrat@87 78086a28-8538-0410-bde1-fa2ac3882955

+++ engines/TorrentPeersModel.cpp
@@ -46,7 +46,7 @@ QVariant TorrentPeersModel::headerData(int section, Qt::Orientation orientation,
 
 QVariant TorrentPeersModel::data(const QModelIndex &index, int role) const
 {
-	if(index.row() >= m_peers.size())
+	if(index.row() >= (int) m_peers.size())
 		return QVariant();
 	const libtorrent::peer_info& info = m_peers[index.row()];
 		
@@ -109,7 +109,7 @@ QVariant TorrentPeersModel::data(const QModelIndex &index, int role) const
 			case 9:
 			{
 				int pcs = 0;
-				for(int i=0;i<info.pieces.size();i++)
+				for(size_t i=0;i<info.pieces.size();i++)
 					if(info.pieces[i])
 						pcs++;
 				return QString("%1%").arg((int) (100.0/double(info.pieces.size())*pcs));
